Rename blobhash utility file.
