+++ fillers/eips/eip4844/test_blobhash_opcode.py
@@ -20,7 +20,7 @@ from ethereum_test_tools import (
 )
 from ethereum_test_tools.vm.opcode import Opcodes as Op
 
-from .blobhash_util import (
+from .util_blobhash import (
     BLOBHASH_GAS_COST,
     MAX_BLOB_PER_BLOCK,
     TARGET_BLOB_PER_BLOCK,
